修改流水金额为int类型，增加查询流水接口

+++ javava-pay/src/main/java/com/javava/pay/entity/FlowRecord.java
@@ -21,7 +21,7 @@ public class FlowRecord {
      * 消费金额
      */
     @Column(name="MONEY")
-    private Double money;
+    private Integer money;
     /**
      * 消费类型 消费/充值
      */
